pyproject.toml
[project]
name = "real-time-visual-defect-detection"
version = "0.1.0"
description = "Thesis pipeline: real-time visual defect detection with robustness evaluation"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "numpy",
  "pyyaml",
  "opencv-python",
  "scikit-learn",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
